refactor: remove db calls from the AI layer
We want database calls to only be made in the message handling layers
and not it the ai layer. The AI layer should only contain the necessary
logic for talking to the LLM and the context needed for the LLM to
respond to a specific audience. This is step 1 of the refactor that will
allow us to break the twitch bot and discord bots into 2 disctinct apps.

Signed-off-by: soypete <petersonmiriah@gmail.com>